First (raw, non-debugged!!!) version of the event app.
models.py now has a first event model class, which can be used to interface
with the site database. Basic event validation is included, which prevents
the user/admin from entering events with inconsistent dates, strips the
state or province if the event is located outside the US or Canada.
Among improvements that could be immediately useful would be setting up a
locations database, which sums up approved locations, has a list of alternative
spellings, zip codes, common mispellings for events. It might help in the long
run with event de-duplication.

admin.py has preliminary rules for displaying/updating the event data.

views.py has the skeleton of the answer to a query for events.