Added category and location models for easier definition of events.
Adding the Location model means that events will always be tied to a
pre-approved location in the database, which should make searching for them
hopefully easier.
Likewise, the Category model was added with possible extentions to subcategories
later on. Now the subcategory is already built into the model, and when we
want to start refining classifications we can start thinking about using this
feature of the Category model.