Adding events, displaying and filtering them appears to work.
It is now possible to enter events from the admin panel, display them,
filter them by date, category and location. Events do not appear to show up
as false duplicates when selecting "All" in the filter list.

NOTE: the event addition has not been properly thoroughly tested yet. In
particular, the Event.clean() method has not been properly debugged.
Some features are still missing or inadequate, for example, the "filter by
date" option currently searches in the past, while it should show options
to filter by future date.
The display mechanism for the "categories" field in the Events display list
causes database calls, so it would be a good idea to switch to another display
mechanism at some point in the future.