Added a timezone field to the location class.
Location is not aware of its time zone, which will prevent problems in the
future where there is uncertainty as to the location of the request. Events
should always be defined in terms of local time, so now any date or time
validation will take place using local time information.