First prototype of a "crawler" which monitors a Google Docs form.
The crawler can read from a google docs form used to manually enter events.
It can build an event structure from this events worksheet, and events can
then be entered and saved in the database. The event creation has to be done
in two steps, first with the creation of a basic event structure, and then
the categorization of the event.