Added read/write event ID from spreadsheet functionality.
It is now possible to read and or write the event ID from the spreadsheet.
If the event ID is empty, it means the event hasn't been added to the DB yet.
Events that do not have IDs are therefore considered to be new events, and
it is now also possible to request a list of new event IDs.