Crawler now looks for updated events on top of new events.
It is now possible to flag updated events in the Google Docs spreadsheet.
The GDocsCrawler will parse this information and use it to build a list of
updated events which should be changed in the database (but not added as new
events).
Also fixed a minor typo in one of the docstrings of models.py